+++ src/components/BounceActivityContainer/ActivityFilter/index.scss
