+++ src/components/BounceRulesContainer/Modals/CreateRuleModal/index.scss
