+++ src/components/BounceRulesContainer/Modals/DeleteRuleModal/index.jsx
@@ -1,10 +1,10 @@
 import React from "react";
 import Button from "@sendgrid/ui-components/button";
 import CenterModal from "@sendgrid/ui-components/center-modal";
-import { Column } from "../../Column";
-import { Row } from "../../Row";
-import "../index.scss";
-import { WriteSelectors } from "../selectors";
+import { Column } from "../../../Column";
+import { Row } from "../../../Row";
+import "../../index.scss";
+import { WriteSelectors } from "../../selectors";
 
 const DeleteAlertBody = () => (
   <div>
